Move the seeder's remote image fetching into a dev-only class
